# Avoid expensive deepcopy when rotating Miller