@hakonanes 2024-12-10
Seems simple enough. I'll have a look within tomorrow.