@hakonanes 2024-12-15
Looking beyond this PR, I think it may be better to avoid `deepcopy()` altogether and rather add a `copy()` method, like this

```python
def copy(self) -> Self:
    return self.__class__(self)
```

We can update the creation method to allow another instance as the first parameter, and only copy the things we need.

We can deprecate `deepcopy()` in favor of `copy()`.

Benefits:
* More control
* New copy via `Vector3d(Vector3d)` or `Vector3d.copy()`

What do you think, @viljarjf?