@hakonanes 2024-12-15
Let's continue the copy discussion in #535.