@hakonanes 2025-01-05
@viljarjf, newly released 0.13.3 has this change in.